Do not run continuous integration on main branch
If a commit is added to the main branch, it should have already run
through the continuous integration pipeline, either on the integration
branch, or via manual trigger in the GitHub UI.

This way we're not running the continuous integration pipeline **both**
before and after a commit is added to the main branch. Only before.